Refactoring of Profile components to be used in Public Profile
