+++ src/web/components/Profile/PublicProfile.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Loading from './Loading';
-import Error from './Error';
+import Loading from '../Loading';
+import Error from '../Error';
 
 const PublicProfile = ({
   loading,
   error,
   firstName,
-  personalities,
+  userPersonalities,
 }) => {
   if (loading) return <Loading />;
 
@@ -17,7 +17,7 @@ const PublicProfile = ({
   return (
     <div>
       <h1>{firstName}</h1>
-      {personalities.map(item => (
+      {userPersonalities.map(item => (
         <div key={item.id}>
           <span>{item.name}</span>
         </div>
@@ -31,7 +31,7 @@ PublicProfile.propTypes = {
   error: PropTypes.string,
   loading: PropTypes.bool.isRequired,
   firstName: PropTypes.string.isRequired,
-  personalities: PropTypes.arrayOf(PropTypes.shape()).isRequired,
+  userPersonalities: PropTypes.arrayOf(PropTypes.shape()).isRequired,
 };
 
 PublicProfile.defaultProps = {
